Moved textual merge related modules to folder and added CSDiff and ConsecutiveLines classes
