+++ src/main/java/br/ufpe/cin/app/JFSTMerge.java
@@ -9,13 +9,13 @@ import br.ufpe.cin.files.FilesEncoding;
 import br.ufpe.cin.files.FilesManager;
 import br.ufpe.cin.files.FilesTuple;
 import br.ufpe.cin.logging.LoggerFactory;
-import br.ufpe.cin.mergers.Diff3;
 import br.ufpe.cin.mergers.SemistructuredMerge;
-import br.ufpe.cin.mergers.TextualMergeStrategy;
+import br.ufpe.cin.mergers.textual.Diff3;
 import br.ufpe.cin.mergers.util.MergeConflict;
 import br.ufpe.cin.mergers.util.MergeContext;
 import br.ufpe.cin.mergers.util.MergeScenario;
 import br.ufpe.cin.mergers.util.RenamingStrategy;
+import br.ufpe.cin.mergers.util.TextualMergeStrategy;
 import br.ufpe.cin.mergers.util.converters.TextualMergeStrategyConverter;
 import br.ufpe.cin.mergers.util.converters.RenamingStrategyConverter;
 import br.ufpe.cin.printers.Prettyprinter;
