+++ src/main/java/br/ufpe/cin/mergers/textual/ConsecutiveLines.java
@@ -1,0 +1,10 @@
+package br.ufpe.cin.mergers.textual;
+
+import br.ufpe.cin.exceptions.TextualMergeException;
+import br.ufpe.cin.mergers.util.TextualMergeStrategy;
+
+public class ConsecutiveLines implements TextualMergeStrategy {
+    public String merge(String leftContent, String baseContent, String rightContent, boolean ignoreWhiteSpaces) throws TextualMergeException {
+        return "";
+    }
+}
